[ACUDIA-28] Jobs and popularity mocked data

+++ lambdas/utils/utils.ts
@@ -1,7 +1,7 @@
 export const random = (min: number, max: number, decimals?: boolean) => {
   const rand = Math.random() * (max - min) + min;
   if (!decimals) return Math.round(rand);
-  return parseInt(rand.toFixed(2));
+  return parseFloat(rand.toFixed(2));
 };
 
 export const formatTimeString = (time: number = 0) => {
